requirements.txt
attrs==20.3.0
awscli==1.18.220
awsebcli==3.19.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
blessed==1.17.12
boto3==1.16.43
botocore==1.19.60
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.1
defusedxml==0.6.0
Django==2.2.17
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-imagekit==4.0.2
django-storages==1.10
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.10
importlib-metadata==2.1.1
jmespath==0.10.0
jsonschema==3.2.0
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
pilkit==2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.4
semantic-version==2.5.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.4.0
